+++ app/src/main/java/com/infomaniak/drive/utils/SentryGridLayoutManager.kt
@@ -19,32 +19,21 @@ package com.infomaniak.drive.utils
 
 import android.content.Context
 import androidx.navigation.NavController
-import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import io.sentry.Sentry
-import io.sentry.SentryLevel
+import com.infomaniak.drive.utils.SentryLinearLayoutManager.Companion.layoutManagerSentryLog
 
-class CustomLinearLayoutManager(private val navController: NavController, context: Context?) : LinearLayoutManager(context) {
+/**
+ * TODO Temp fix
+ */
+class SentryGridLayoutManager(private val navController: NavController, context: Context?, spanCount: Int) :
+    GridLayoutManager(context, spanCount) {
 
     override fun onLayoutChildren(recycler: RecyclerView.Recycler?, state: RecyclerView.State?) {
         try {
             super.onLayoutChildren(recycler, state)
         } catch (exception: IndexOutOfBoundsException) {
             layoutManagerSentryLog(navController, exception)
-        }
-    }
-
-    companion object {
-        fun layoutManagerSentryLog(navController: NavController, exception: IndexOutOfBoundsException) {
-            exception.printStackTrace()
-            Sentry.withScope { scope ->
-                navController.currentDestination?.displayName?.let { name ->
-                    scope.setExtra("navigation", name)
-                }
-                scope.level = SentryLevel.WARNING
-                scope.setExtra("message", "Data modified in different thread")
-                Sentry.captureException(exception)
-            }
         }
     }
 }
